Correct license headers introduced by Arm Limited
This commit removes the additional clause
"THIS HEADER MAY NOT BE EXTRACTED OR MODIFIED IN ANY WAY"
from BSD license headers that got introduced by mistake
with the following commits:
 86a47bf - external plat: Add platform interface implementation (Haibo Xu)
 fa5c59f - external plat: Add the solo5 API declaration (Haibo Xu)

On the one hand, this additional clause is redundant because
the BSD license already states that it must remain and the
copyright notice must be kept.
On the other hand, the clause freezes the file header
and prohibits future contributors from extending the
copyright notice for their contributions. This additional
clause is not part of the official BSD 3-clause.

The original author(s) or an authorized representative from
the author's affiliation consents to the change with a
`Reviewed-by` tag to this commit.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Jia He <justin.he@arm.com>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Github-Pull-Request: #2